eol_program_test/coscoot_diag_tool/requirements.txt
PyQt5~=5.15.4
pyserial~=3.4
PyQt5-stubs
pyuic5-tool
python-can~=3.3.4
requests
crcmod
pyqtdarktheme
